TeEditorBase
