+++ src/TableViewerBlazor/Internal/TeComponent/TeEditorBase.cs
@@ -1,0 +1,9 @@
+namespace TableViewerBlazor.Internal.TeComponent;
+
+public class TeEditorBase : ComponentBase
+{
+    [Parameter] public object? Data { get; set; }
+    [Parameter] public EventCallback<object?> DataChanged { get; set; }
+    [Parameter] public TeOptions Options { get; set; } = default!;
+    [Parameter] public int Depth { get; set; }
+}
