+++ src/TableViewerBlazor/Internal/TeComponent/TeSelectBox.razor
@@ -1,3 +1,5 @@
+@inherits TeEditorBase
+
 <MudSelect
     T="object"
     Value="Data"
