+++ src/TableViewerBlazor/TableEditor.razor
@@ -3,4 +3,6 @@
 <TeElementEditor Type="@typeof(T)"
     Data="Data"
     DataChanged="@(async value => await OnDataChanged(value))"
-    Options="Options"/>
+    Options="Options"
+    Depth="1"
+    />
