+++ test/TableViewerTest/Components/Pages/EditorPage.razor
@@ -4,3 +4,5 @@
     Data="data"
     Options="options"
     DataChanged="Changed" />
+
+<TableViewer Data="viewData" />
