+++ test/TableViewerTest/Components/Pages/EditorPage.razor.cs
@@ -9,6 +9,7 @@ public class EditData
     public string Name { get; set; } = string.Empty;
     [TeSelectBox("gender")]
     public string Gender { get; set; } = string.Empty;
+    [TeSelectBox("age")]
     public int Age { get; set; }
 }
 
@@ -22,6 +23,7 @@ public partial class EditorPage : ComponentBase
         Gender = "M",
         Age = 30,
     };
+    EditData? viewData;
     TeOptions options = new TeOptions
     {
         SelectBoxOptions = new ITeSelectBoxOption[]
@@ -31,14 +33,27 @@ public partial class EditorPage : ComponentBase
                 Id = "gender",
                 Items = new TeSelectItem<string>[]
                 {
-                    new TeSelectItem<string>("남자", "M"),
-                    new TeSelectItem<string>("여자", "F"),
+                    new TeSelectItem<string>("M", "남자"),
+                    new TeSelectItem<string>("F", "여자"),
+                },
+            },
+            new TeSelectBoxOption<int>
+            {
+                Id = "age",
+                Items = new TeSelectItem<int>[]
+                {
+                    new TeSelectItem<int>(10, "10대"),
+                    new TeSelectItem<int>(20, "20대"),
+                    new TeSelectItem<int>(30, "30대"),
+                    new TeSelectItem<int>(40, "40대"),
+                    new TeSelectItem<int>(50, "50대"),
                 },
             },
         },
     };
     private async Task Changed(EditData data)
     {
+        viewData = data;
         await Js.InvokeVoidAsync("console.log", data);
     }
 }
